Data.MultiSet

Description

An efficient implementation of multisets, also sometimes called bags.

A multiset is like a set, but it can contain multiple copies of the same element. Unless otherwise specified all insert and remove opertions affect only a single copy of an element. For example the minimal element before and after deleteMin could be the same, only with one less occurrence.

Since many function names (but not the type name) clash with Prelude names, this module is usually imported qualified, e.g.

 import Data.MultiSet (MultiSet)
 import qualified Data.MultiSet as MultiSet

The implementation of MultiSet is based on the Data.Map module.

Note that the implementation is left-biased -- the elements of a first argument are always preferred to the second, for example in union or insert. Of course, left-biasing can only be observed when equality is an equivalence relation instead of structural equality.

In the complexity of functions n refers to the number of distinct elements, t is the total number of elements.

MultiSet type

data MultiSet a

A multiset of values a. The same value can occur multiple times.

type Occur = Int

The number of occurrences of an element

Operators

(\\) :: Ord a => MultiSet a -> MultiSet a -> MultiSet a infixl 9

O(n+m). See difference.

Query

null :: MultiSet a -> Bool

O(1). Is this the empty multiset?

size :: MultiSet a -> Occur

O(n). The number of elements in the multiset.

distinctSize :: MultiSet a -> Occur

O(1). The number of distinct elements in the multiset.

member :: Ord a => a -> MultiSet a -> Bool

O(log n). Is the element in the multiset?

notMember :: Ord a => a -> MultiSet a -> Bool

O(log n). Is the element not in the multiset?

occur :: Ord a => a -> MultiSet a -> Occur

O(log n). The number of occurrences of an element in a multiset.

isSubsetOf :: Ord a => MultiSet a -> MultiSet a -> Bool

O(n+m). Is this a subset? (s1 `isSubsetOf` s2) tells whether s1 is a subset of s2.

isProperSubsetOf :: Ord a => MultiSet a -> MultiSet a -> Bool

O(n+m). Is this a proper subset? (ie. a subset but not equal).

Construction

empty :: MultiSet a

O(1). The empty mutli set.

singleton :: a -> MultiSet a

O(1). Create a singleton mutli set.

insert :: Ord a => a -> MultiSet a -> MultiSet a

O(log n). Insert an element in a multiset.

insertMany :: Ord a => a -> Occur -> MultiSet a -> MultiSet a

O(log n). Insert an element in a multiset a given number of times.

Negative numbers remove occurrences of the given element.

delete :: Ord a => a -> MultiSet a -> MultiSet a

O(log n). Delete a single element from a multiset.

deleteMany :: Ord a => a -> Occur -> MultiSet a -> MultiSet a

O(log n). Delete an element from a multiset a given number of times.

Negative numbers add occurrences of the given element.

deleteAll :: Ord a => a -> MultiSet a -> MultiSet a

O(log n). Delete all occurrences of an element from a multiset.

Combine

union :: Ord a => MultiSet a -> MultiSet a -> MultiSet a

O(n+m). The union of two multisets. The union adds the occurrences together.

The implementation uses the efficient hedge-union algorithm. Hedge-union is more efficient on (bigset union smallset).

unions :: Ord a => [MultiSet a] -> MultiSet a

The union of a list of multisets: (unions == foldl union empty).

maxUnion :: Ord a => MultiSet a -> MultiSet a -> MultiSet a

O(n+m). The union of two multisets. The number of occurrences of each element in the union is the maximum of the number of occurrences in the arguments (instead of the sum).

The implementation uses the efficient hedge-union algorithm. Hedge-union is more efficient on (bigset union smallset).

difference :: Ord a => MultiSet a -> MultiSet a -> MultiSet a

O(n+m). Difference of two multisets. The implementation uses an efficient hedge algorithm comparable with hedge-union.

intersection :: Ord a => MultiSet a -> MultiSet a -> MultiSet a

O(n+m). The intersection of two multisets. Elements of the result come from the first multiset, so for example

import qualified Data.MultiSet as MS
data AB = A | B deriving Show
instance Ord AB where compare _ _ = EQ
instance Eq AB where _ == _ = True
main = print (MS.singleton A `MS.intersection` MS.singleton B,
              MS.singleton B `MS.intersection` MS.singleton A)

prints (fromList [A],fromList [B]).

Filter

filter :: (a -> Bool) -> MultiSet a -> MultiSet a

O(n). Filter all elements that satisfy the predicate.

partition :: (a -> Bool) -> MultiSet a -> (MultiSet a, MultiSet a)

O(n). Partition the multiset into two multisets, one with all elements that satisfy the predicate and one with all elements that don't satisfy the predicate. See also split.

split :: Ord a => a -> MultiSet a -> (MultiSet a, MultiSet a)

O(log n). The expression (split x set) is a pair (set1,set2) where all elements in set1 are lower than x and all elements in set2 larger than x. x is not found in neither set1 nor set2.

splitOccur :: Ord a => a -> MultiSet a -> (MultiSet a, Occur, MultiSet a)

O(log n). Performs a split but also returns the number of occurrences of the pivot element in the original set.

Map

map :: Ord b => (a -> b) -> MultiSet a -> MultiSet b

O(n*log n). map f s is the multiset obtained by applying f to each element of s.

mapMonotonic :: (a -> b) -> MultiSet a -> MultiSet b

O(n). The

mapMonotonic f s == map f s, but works only when f is strictly monotonic. The precondition is not checked. Semi-formally, we have:

and [x < y ==> f x < f y | x <- ls, y <- ls]
                    ==> mapMonotonic f s == map f s
    where ls = toList s

mapMaybe :: Ord b => (a -> Maybe b) -> MultiSet a -> MultiSet b

O(n). Map and collect the Just results.

mapEither :: (Ord b, Ord c) => (a -> Either b c) -> MultiSet a -> (MultiSet b, MultiSet c)

O(n). Map and separate the Left and Right results.

concatMap :: Ord b => (a -> [b]) -> MultiSet a -> MultiSet b

O(n). Apply a function to each element, and take the union of the results

unionsMap :: Ord b => (a -> MultiSet b) -> MultiSet a -> MultiSet b

O(n). Apply a function to each element, and take the union of the results

Monadic

bind :: Ord b => MultiSet a -> (a -> MultiSet b) -> MultiSet b

O(n). The monad bind operation, (>>=), for multisets.

join :: Ord a => MultiSet (MultiSet a) -> MultiSet a

O(n). The monad join operation for multisets.

Fold

fold :: (a -> b -> b) -> b -> MultiSet a -> b

O(t). Fold over the elements of a multiset in an unspecified order.

foldOccur :: (a -> Occur -> b -> b) -> b -> MultiSet a -> b

O(n). Fold over the elements of a multiset with their occurrences.

Min/Max

findMin :: MultiSet a -> a

O(log n). The minimal element of a multiset.

findMax :: MultiSet a -> a

O(log n). The maximal element of a multiset.

deleteMin :: MultiSet a -> MultiSet a

O(log n). Delete the minimal element.

deleteMax :: MultiSet a -> MultiSet a

O(log n). Delete the maximal element.

deleteMinAll :: MultiSet a -> MultiSet a

O(log n). Delete all occurrences of the minimal element.

deleteMaxAll :: MultiSet a -> MultiSet a

O(log n). Delete all occurrences of the maximal element.

deleteFindMin :: MultiSet a -> (a, MultiSet a)

O(log n). Delete and find the minimal element.

deleteFindMin set = (findMin set, deleteMin set)

deleteFindMax :: MultiSet a -> (a, MultiSet a)

O(log n). Delete and find the maximal element.

deleteFindMax set = (findMax set, deleteMax set)

maxView :: MultiSet a -> Maybe (a, MultiSet a)

O(log n). Retrieves the maximal element of the multiset, and the set with that element removed. Returns Nothing when passed an empty multiset.

Examples:

>>> maxView $ fromList ['a', 'a', 'b', 'c']
Just ('c',fromOccurList [('a',2),('b',1)])

minView :: MultiSet a -> Maybe (a, MultiSet a)

O(log n). Retrieves the minimal element of the multiset, and the set with that element removed. Returns Nothing when passed an empty multiset.

Examples:

>>> minView $ fromList ['a', 'a', 'b', 'c']
Just ('a',fromOccurList [('a',1),('b',1),('c',1)])

Conversion

List

elems :: MultiSet a -> [a]

O(t). The elements of a multiset.

distinctElems :: MultiSet a -> [a]

O(n). The distinct elements of a multiset, each element occurs only once in the list.

distinctElems = map fst . toOccurList

toList :: MultiSet a -> [a]

O(t). Convert the multiset to a list of elements.

fromList :: Ord a => [a] -> MultiSet a

O(t*log t). Create a multiset from a list of elements.

Ordered list

toAscList :: MultiSet a -> [a]

O(t). Convert the multiset to an ascending list of elements.

fromAscList :: Eq a => [a] -> MultiSet a

O(t). Build a multiset from an ascending list in linear time. The precondition (input list is ascending) is not checked.

fromDistinctAscList :: [a] -> MultiSet a

O(n). Build a multiset from an ascending list of distinct elements in linear time. The precondition (input list is strictly ascending) is not checked.

Occurrence lists

toOccurList :: MultiSet a -> [(a, Occur)]

O(n). Convert the multiset to a list of element/occurrence pairs.

toAscOccurList :: MultiSet a -> [(a, Occur)]

O(n). Convert the multiset to an ascending list of element/occurrence pairs.

fromOccurList :: Ord a => [(a, Occur)] -> MultiSet a

O(n*log n). Create a multiset from a list of element/occurrence pairs. Occurrences must be positive. The precondition (all occurrences > 0) is not checked.

fromAscOccurList :: Eq a => [(a, Occur)] -> MultiSet a

O(n). Build a multiset from an ascending list of element/occurrence pairs in linear time. Occurrences must be positive. The precondition (input list is ascending, all occurrences > 0) is not checked.

fromDistinctAscOccurList :: [(a, Occur)] -> MultiSet a

O(n). Build a multiset from an ascending list of elements/occurrence pairs where each elements appears only once. Occurrences must be positive. The precondition (input list is strictly ascending, all occurrences > 0) is not checked.

Map

toMap :: MultiSet a -> Map a Occur

O(1). Convert a multiset to a Map from elements to number of occurrences.

fromMap :: Map a Occur -> MultiSet a

O(n). Convert a Map from elements to occurrences to a multiset.

fromOccurMap :: Map a Occur -> MultiSet a

O(1). Convert a Map from elements to occurrences to a multiset. Assumes that the Map contains only values larger than zero. The precondition (all elements > 0) is not checked.

Set

toSet :: MultiSet a -> Set a

O(n). Convert a multiset to a Set, removing duplicates.

fromSet :: Set a -> MultiSet a

O(n). Convert a Set to a multiset.

Debugging

showTree :: Show a => MultiSet a -> String

O(n). Show the tree that implements the set. The tree is shown in a compressed, hanging format.

showTreeWith :: Show a => Bool -> Bool -> MultiSet a -> String

O(n). The expression (showTreeWith hang wide map) shows the tree that implements the set. If hang is True, a hanging tree is shown otherwise a rotated tree is shown. If wide is True, an extra wide version is shown.

Set> putStrLn $ showTreeWith True False $ fromDistinctAscList [1,1,2,3,4,5]
(1*) 4
+--(1*) 2
|  +--(2*) 1
|  +--(1*) 3
+--(1*) 5

Set> putStrLn $ showTreeWith True True $ fromDistinctAscList [1,1,2,3,4,5]
(1*) 4
|
+--(1*) 2
|  |
|  +--(2*) 1
|  |
|  +--(1*) 3
|
+--(1*) 5

Set> putStrLn $ showTreeWith False True $ fromDistinctAscList [1,1,2,3,4,5]
+--(1*) 5
|
(1*) 4
|
|  +--(1*) 3
|  |
+--(1*) 2
   |
   +--(2*) 1

valid :: Ord a => MultiSet a -> Bool

O(n). Test if the internal multiset structure is valid.